wip refactoring chain sync to workers in async tasks
