+++ forest/src/daemon.rs
@@ -93,7 +93,7 @@ pub(super) async fn start(config: Config) {
     let bad_blocks = chain_syncer.bad_blocks_cloned();
     let sync_state = chain_syncer.sync_state_cloned();
     let sync_task = task::spawn(async {
-        chain_syncer.start().await.unwrap();
+        chain_syncer.start().await;
     });
 
     // Start services
